utils: move all URL methods to utils.url
