+++ src/streamlink/plugins/abweb.py
@@ -4,8 +4,7 @@ import re
 from streamlink.plugin import Plugin, PluginArgument, PluginArguments, PluginError, pluginmatcher
 from streamlink.plugin.api.utils import itertags
 from streamlink.stream import HLSStream
-from streamlink.utils import update_scheme
-from streamlink.utils.url import url_concat
+from streamlink.utils.url import update_scheme, url_concat
 
 log = logging.getLogger(__name__)
