+++ src/streamlink/plugins/ard_mediathek.py
@@ -4,7 +4,7 @@ import re
 from streamlink.plugin import Plugin, pluginmatcher
 from streamlink.plugin.api import validate
 from streamlink.stream import HLSStream, HTTPStream
-from streamlink.utils import update_scheme
+from streamlink.utils.url import update_scheme
 
 MEDIA_URL = "http://www.ardmediathek.de/play/media/{0}"
 QUALITY_MAP = {
