+++ src/streamlink/plugins/atresplayer.py
@@ -4,7 +4,8 @@ import re
 from streamlink.plugin import Plugin, pluginmatcher
 from streamlink.plugin.api import validate
 from streamlink.stream import DASHStream, HLSStream
-from streamlink.utils import parse_json, search_dict, update_scheme
+from streamlink.utils import parse_json, search_dict
+from streamlink.utils.url import update_scheme
 
 log = logging.getLogger(__name__)
