+++ src/streamlink/plugins/mediaklikk.py
@@ -5,7 +5,8 @@ from urllib.parse import unquote, urlparse
 from streamlink.plugin import Plugin, pluginmatcher
 from streamlink.plugin.api import validate
 from streamlink.stream import HLSStream
-from streamlink.utils import parse_json, update_scheme
+from streamlink.utils import parse_json
+from streamlink.utils.url import update_scheme
 
 
 log = logging.getLogger(__name__)
