+++ src/streamlink/plugins/webtv.py
@@ -8,8 +8,9 @@ from Crypto.Cipher import AES
 from streamlink.plugin import Plugin, pluginmatcher
 from streamlink.plugin.api import validate
 from streamlink.stream import HLSStream
-from streamlink.utils import parse_json, update_scheme
+from streamlink.utils import parse_json
 from streamlink.utils.crypto import unpad_pkcs5
+from streamlink.utils.url import update_scheme
 
 log = logging.getLogger(__name__)
