+++ tests/test_utils_url.py
@@ -1,7 +1,26 @@
 from collections import OrderedDict
 from urllib.parse import quote
 
-from streamlink.utils.url import update_qsd, update_scheme, url_concat, url_equal
+import pytest
+
+from streamlink.utils.url import absolute_url, prepend_www, update_qsd, update_scheme, url_concat, url_equal
+
+
+@pytest.mark.parametrize("baseurl,url,expected", [
+    ("http://test.se", "/test", "http://test.se/test"),
+    ("http://test.se", "http/test.se/test", "http://test.se/http/test.se/test"),
+    ("http://test.se", "http://test2.se/test", "http://test2.se/test"),
+])
+def test_absolute_url(baseurl, url, expected):
+    assert expected == absolute_url(baseurl, url)
+
+
+@pytest.mark.parametrize("url,expected", [
+    ("http://test.se/test", "http://www.test.se/test"),
+    ("http://www.test.se", "http://www.test.se"),
+])
+def test_prepend_www(url, expected):
+    assert expected == prepend_www(url)
 
 
 def test_update_scheme():
